// marl-ui/src/types/config.d.ts
export interface GlobConfig {
  // Site title
  title: string
  menuTitle: string
  // Service interface url
  // apiUrl: string;
  systemApiUrl?: string
  productApiUrl?: string
  categoryApiUrl?: string
  certificateApiUrl?: string
  userApiUrl?: string
  appApiUrl?: string
  oAuth2Url?: string
  // Upload url
  uploadUrl?: string
  //  Service interface url prefix
  urlPrefix?: string
  // Project abbreviation
  shortName: string
}